Mini fixes

+++ benchmarks/atoi.c
@@ -2,9 +2,9 @@
 #include <string.h>
 
 /* We have to wrap atoi because it expects a null-terminated string */
-int hs_atoi(char *str, int len) {
+long hs_atoi(char *str, int len) {
     char *str0 = malloc(len + 1);
-    int x;
+    long x;
     strncpy(str0, str, len);
     str0[len] = '\0';
     x = atoi(str0);
